Restore compatibility with latest `@yowasp/runtime`.
The packages now generate a statement of the form `() => import`, which
was not handled by the hacky loader that the desktop code paths use.

Replacing `/\bimport\b/g` breaks pyodide. It is probably viable to
replace `/\bimport\(/g`, but that might still break things, and right
now loading the latest yosys bundle crashes the extension host, so
this commit does the simplest possible thing.

I can't wait to remove `importModuleCriminally`, which can be done once
microsoft/vscode#201935 is done.